+++ src/app/main/map/map-view/map-view.component.ts
@@ -114,7 +114,7 @@ export class MapViewComponent implements OnInit, AfterViewInit, OnDestroy {
     this.metaService.setTagsGeneral(
       this.translateService.instant('main.map-view.title_page'),
       [
-        new MetaTag(METATAG.KEYWORDS, 'njuka, coupure electricité, coupure eau, délestage, Eneo, Camwater, Cameroun, Cameroon, signaler coupure, panne de courant, coupures programmées, power outage, water outage, carte des coupures'),
+        new MetaTag(METATAG.KEYWORDS, 'njuka, coupure electricité, coupure eau, délestage, SOCADEL, Eneo, Camwater, Cameroun, Cameroon, signaler coupure, panne de courant, coupures programmées, power outage, water outage, carte des coupures'),
         new MetaTag(METATAG.DESCRIPTION, this.translateService.instant('main.map-view.desc_page'))
       ]);
   }