+++ src/app/main/static-page/about-us/about-us.component.ts
@@ -36,7 +36,7 @@ export class AboutUsComponent implements OnInit {
     this.metaService.setTagsGeneral(
       this.translateService.instant('core.aboutus.title_page'),
       [
-      new MetaTag(METATAG.KEYWORDS, 'njuka, coupure electricité, coupure eau, délestage, Eneo, Camwater, Cameroun, Cameroon, signaler coupure, panne de courant, coupures programmées, power outage, water outage, carte des coupures'),
+      new MetaTag(METATAG.KEYWORDS, 'njuka, coupure electricité, coupure eau, délestage, SOCADEL, Eneo, Camwater, Cameroun, Cameroon, signaler coupure, panne de courant, coupures programmées, power outage, water outage, carte des coupures'),
       new MetaTag(METATAG.DESCRIPTION, this.translateService.instant('core.aboutus.desc_page'))
     ]);
   }